Add txData controller
